app/serializers/Pokemon.py
from rest_framework import serializers

from app.models.Pokemon import Pokemon

class PokemonSerializer(serializers.ModelSerializer):
    class Meta:
        model = Pokemon
        fields = '__all__'
