requirements.txt
Django>=3.0,<4.0
psycopg2-binary>=2.8
djangorestframework==3.11.2
requests==2.28.1
pytest==7.1.2